backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyMuPDF
google-generativeai
python-multipart
aiofiles
reportlab